requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
backports.zoneinfo==0.2.1
    # via django
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==40.0.2
    # via jwcrypto
decorator==5.1.1
    # via ipython
deprecated==1.2.13
    # via jwcrypto
django==4.2.1
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
    #   djangorestframework
django-oauth-toolkit==2.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
idna==3.4
    # via requests
ipython==8.12.2
    # via -r requirements.in
jedi==0.18.2
    # via ipython
jwcrypto==1.4.2
    # via django-oauth-toolkit
matplotlib-inline==0.1.6
    # via ipython
oauthlib==3.2.2
    # via django-oauth-toolkit
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.38
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via ipython
pytz==2023.3
    # via djangorestframework
requests==2.30.0
    # via django-oauth-toolkit
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via django
stack-data==0.6.2
    # via ipython
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via ipython
urllib3==2.0.2
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
wrapt==1.15.0
    # via deprecated
